requirements.txt
alembic==1.6.5
amqp==5.0.6
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
packaging==21.0
pkg_resources==0.0.0
pluggy==0.13.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
